requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==24.2.0
av==13.1.0
blinker==1.8.2
boto3==1.35.19
botocore==1.35.19
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
compressed-tensors==0.8.0
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.32.1
datasets==3.1.0
Deprecated==1.2.14
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
einops==0.8.0
fastapi==0.115.5
filelock==3.13.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.2.0
gast==0.6.0
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.4.0
interegular==0.3.3
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lark==1.2.2
libclang==18.1.1
llvmlite==0.43.0
lm-format-enforcer==0.10.9
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mistral_common==1.5.1
ml-dtypes==0.4.0
mlflow==2.16.1
mlflow-skinny==2.16.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvidia-ml-py==12.560.30
openai==1.54.3
opencv-python-headless==4.10.0.84
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt-einsum==3.3.0
optree==0.12.1
outlines==0.0.46
packaging==24.1
pandas==2.2.2
partial-json-parser==0.2.1.1.post4
pillow==10.4.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
propcache==0.2.0
protobuf==4.25.4
psutil==6.1.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
qwen-vl-utils==0.0.8
ray==2.39.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.21.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.3.1
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.34
sqlparse==0.5.1
starlette==0.41.3
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tk==0.1.0
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.4.1+cu121
torchinfo==1.8.0
torchvision==0.20.1
tqdm==4.66.5
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.32.1
vllm==0.6.4.post1
waitress==3.0.0
watchfiles==1.0.0
websockets==14.1
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.18.0
zipp==3.20.2